Fix regression with db_setup.
When you run db_setup with Bundler, it sets some environment settings that we
don't want set (since db_setup uses bundler itself, which retains those
incorrect settings). This was causing db_setup to do nothing, which is a
problem if we ever need to run migrations. Fixed the issue by unsetting
the conflicting environment settings.