Start of the code refactor.
The previous code base was problematic for many reasons. The biggest one
is that it consisted of three modules that stacked on top of each other:
runner, master and frontend.

The end goal of the refactor is to break Buffet into discrete modules
(particularly, modules that don't have interdependencies). This is
helpful if, say, we want to swap in a new way to setup the working
directory, or a test distributer that isn't RSpec specific.

The new modules are:

  * Setup
    This module brings the working directory into a testable state.

  * Master
    This module is the test distributer. It runs the worker module on
    each host.

  * Worker
    Simple module. While tests haven't been tested, run one of them.

  * Buffet
    The core module of Buffet. It uses Setup and Master, and also
    exposes some helpful test-related information.

  * Frontend
    The web interface to Buffet.

  * CLI
    The command line interface to Buffet.