Fix status output error.
Stemmed from the fact that implementing to_s on an object makes the
object look like a string, but not play like one. So if you call string
methods on the object, it doesn't work. This is tricky. It suggests that
I should probably make StatusMessage inherit from string, or ditch to_s
entirely.