Add failing test case for db_setup.
Particularly, this test case tests for the case where the current
machine is not listed under hosts in settings.yml. In this case,
potentially conflicting operations are done with a machine that is in
settings.yml.

The purpose of this is to move towards being able to run buffet twice
on one computer. The current reason it doesn't work is because you can't
nuke the db while buffet is running, but Buffet isn't smart enough to
nuke the db from a different host.