Don't start Buffet in a thread by default.
The choice of whether Buffet should be in a separate thread is instead
left up to the user of the class.

Apart from being good style, this is needed for the CLI, since it
doesn't make sense for the CLI to run Buffet asynchronously unless asked
(i.e. bundle exec ruby cli.rb &).