Rsync should ignore .bundle.
Before this change, rsync would copy over the .bundle directory. The
problem with this is that it would copy over the config, which would
create obscure errors when it tried to write to nonexistant directories
later.

Now we don't copy over the .bundle directory, so once we set the bundle
settings correctly the first time, we don't need to ever touch them
again.