Generalize buffet_worker.
The old buffet_worker script only worked for one obscure version of
rspec. The new one will be able to work more universally.

The advantage of this approach is that we can essentially integrate any
testing framework with Buffet.

The disadvantage is that using regular expressions to match test
passes/failures is kind of a hack. Right now, we're guaranteed that it
will work because of the way that rspec formats the output. However,
it's not guaranteed that this will always work.

There's no perfect solution to this problem. It's not guaranteed
that every testing framework provides a automateable interface
for us to use, but on the other hand, regular expressions can be
tripped up by wonky output (It's not likely, but it could happen).

The "correct" solution is to build interfaces into every testing
framework that doesn't have them, or to create pull requests on them.
But that's beyond the scope of Buffet, at least for now.