Fix PID_FILE 'lock'.
You can run Buffet twice on the same computer, so long as:

  1) The two instances are not running on the same directory.

  2) There is no intersection in the set of hosts of the runs.
  (Crucially, only one Buffet instance can use the machine that it's
  running on.)

In fact, you can (in theory) run Buffet an arbitrary amount of times on
a single computer, if you want to.

My usecase of this feature is to have 'production' and 'development'
versions of Buffet. There are other reasons it could be used, as well.