Conditionally bundle install on bundle check.
This is one of the biggest slowdowns when running very small test
suites. Sure, that's not the biggest planned use case of Buffet, but
it's helpful when testing Buffet with other small repositories, which
I'll be doing a lot of in the future. It's also really helpful when you test
Buffet with Buffet.

B U F F C E P T...nevermind.