Add regression testing to Buffet.
Currently, regression testing only works with the CLI, and the output
format is suboptimal. Still - it's functional!

A test is considered to be a regression if:

  1. It has failed the most recent test run.

  2. It existed the last test run, and passed the last test run.

Regression testing aims to inform Buffet users whether their particular
commit broke anything, which is potentially more useful then knowing
that x things are broke in general.