Fix RSpec1 regression.
This is one of the stranger bugs I've ever seen. I think it happens because
RSpec1 silently closes the files that you pass to it via the CommandLine
runner.

This is nasty behavior, mainly because it's unexpected and because the
error that you eventually see has little to do with the cause.

They seem to have fixed the problem in RSpec2, which is nice, but not
helpful if you're not working on RSpec2... :-)

I think this was done because the typical use case is not passing in a
file object, but rather a string containing the location of the file.
When you do that, it makes sense that RSpec handles the creation and
deletion of the file, and so RSpec carries that over to passed in file
objects. However, this definitely isn't the right thing to do with file
objects.