Fix run tests button.
Was hearing frequent reports of weird stuff happening with the run tests
question on the top of the test output. Sometimes it would exist when it
shouldn't, and sometimes it didn't exist when it should. This was
because I didn't have an AJAX poller for it like I did for everything
else on Buffet.

Now, the button should exist at precisely the right times.