Fix RSpec1/2 incompatibility in worker.
There is really frustrating inconsistency with how RSpec1 and 2 handle
the file handles that you pass in. I believe that this commit makes them
both work happily.

The issue was that RSpec2 does not flush the list of files that you're
running tests on between tests. This is only an issue when you run many
tests in a row. What it means practically is that if you run tests 1, 2,
3, 4, and 5, it'll run 1, then 1, 2, then 1, 2, 3... etc.

This fixes that issue.