Fix some issues with the frontend.
One problem was that I had moved from storing a hashtable in
status_message to a string, and so when I was trying to read that out as
a hash_table, things were getting confused. This suggests that the use
of status_message is kind of confused, since it stores data across
the different modules and occasionally displays it also.

There was another problem with messages flashing by too quickly on the
web interface. I've since made sure that the most important messages are
visible.