Check-mode checks for bundler correctly.
Had a typo where I was checking "hosts" for bundler, not a particular
"host". This commit fixes that typo.

Also, I was previously checking the exit status of ssh, which will
always be successful unless this machine cannot access the other (which
is not a possible outcome barring races since we've already verified
that we can). Now, we check the actual contents of the output instead.