Fix critical error with db_setup.
db_setup had been evaluating the arguments passed in instead of just
quoting them. This was causing the script to crash before updating
migrations. Since migrations weren't being updated, Buffet would
incorrectly say that many tests that were perfectly fine were failing.

This went undetected for so long because the crash would happen at the
last line of the file, not the line with the syntax error, so it seemed
fairly innocuous ('oh, the error happened after the file ran, so it's
not a big deal').

Furthermore, git blame shows that the line has existed untouched since
the very first Buffet commit (!), so it was assumed that that failing
was just typical behavior and not indicative of serious problems.

Change-Id: I258bad9f21ae522d4cb447c4b7c1e55e5a1ddc88