Fix broken Settings.create_settings_file call
Running the spec suite was exploding trying to call create_settings_file,
defined on the Buffet module instead of the Buffet::Settings class.

With this change the spec suite actually runs, although there are four
failures, and the spec suite drops you into an editor inviting you to configure
Buffet for the first time and proceeds to create a ~/.buffet directory and fill
it with stuff; running the spec suite should be free of that kind of side
effect, so that will need to be fixed in another commit.

Change-Id: I6113ff5c05a06a27c930de78c445428de9fa2892
Reviewed-on: https://gerrit.causes.com/726
Reviewed-by: Lilia Markham <lilia.markham@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>