Remove --watch facility
In the interests of keeping Buffet lean and maintainable, jettison this
feature which is largely duplicative of the CI provided by other tools such as
Jenkins, and which Jenkins is able to do with much more power and flexibility.

In subsequent commits I'll be cutting some of the extra bells and whistles and
trying to reduce Buffet to a tool that does a specific task (running tests in a
parallel fashion) effectively and reliably.

Change-Id: Ifce8cf748008caebb240401a7cbc85fede8e3683
Reviewed-on: https://gerrit.causes.com/728
Reviewed-by: Lilia Markham <lilia.markham@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>