Remove web frontend
Ultimately the web frontend doesn't match our workflow, because we do not want
to kick off a run in a web browser and then check back on that window for the
results; rather, the command-line interface suits us because we can plug it in
to something like Jenkins, and have Jenkins worry about the details of
notifying us of the results (via Campfire, or integration with Gerrit etc).

Removing code which we won't be using frees us from having to maintain and
support it.

Change-Id: Ib2425c5e33813c0f3cb62e049e57ecd941e24ed4
Reviewed-on: https://gerrit.causes.com/732
Reviewed-by: Jimmy Kittiyachavalit <jimmy@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>