Leave management of settings file up to user
Attempting to run the spec suite, or even just run Buffet, should not create a
bunch of files in your home directory. Likewise, if a user wishes to edit an
existing settings file, he/she can just open it in their editor. Removing this
code in the interests of keeping Buffet down to a maintainable minimum.

Note that the spec suite is continues to be in a broken state; I'll aim to get
it green once I've finished stripping things down to a bare minimum.

Change-Id: Ie431173c22deb7075c688dd104709ec3b5d2cc63
Reviewed-on: https://gerrit.causes.com/733
Reviewed-by: Jimmy Kittiyachavalit <jimmy@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>