Large refactor into working minimal-buffet
This is a working version of minimal buffet. Many files have
been removed or refactored where their functionality was deemed
unnecessary.

The gem was also renamed to `minimal-buffet`, and the version set
to 0.6.

We've settled on the "convention over configuration" mantra. For
example, we expect the repo being tested to have a "buffet.yml"
file defining the project and the slaves to use when running it.
We also assume the existence of a script bin/before-buffet-run in
the repo, which is a user-specified set of actions to do (migrations,
etc) before running specs with buffet.

Change-Id: I1c06eca88183b33f05ab9b21abb643872ab9c21b
Reviewed-on: https://gerrit.causes.com/1708
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>