Prevent race condition on DRB server startup
Previously, a race condition could occur where the Master would
attempt to access the DRb server before it was initialized on
the other thread.

This commit fixes this problem by using a condition variable
to wait until the server is initialized before returning from
the start_service call in Master.

Change-Id: I558bc5b76b82733411120df77e9c186896f22441
Reviewed-on: https://gerrit.causes.com/1718
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>