Get remote connections working
Previously, buffet-worker processes could not connect to the
Master because the master was sending the workers its local
all-interface address (0.0.0.0), which meant the wrong address
to the remote workers.

This was fixed by parsing the output of host and hostname to
retrieve the local IP address for the machine on the internal
network.

Change-Id: Ied459eff897aefc246dbd4da58b710dc1482a9df
Reviewed-on: https://gerrit.causes.com/1719
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>