Accommodate projects that don't need a set-up script
Simple projects (for example, projects that don't need a "bundle
install" or a database migration) don't need to run a set-up script at
all, so for projects which neither define a "prepare_script" setting in
their buffet.yml, nor have a file named "bin/before-buffet-run" in their
directory, we don't try to run any set-up script.

To acheive this, added a '.has_prepare_script?' method on
Buffet::Settings. As all the methods in the class are class methods,
moved them inside a "class << self" block to avoid the need for the
repeated "def self.*" pattern.

I added specs for the new method, and the [] method that it relies on,
but not for the other methods in the file. Those can be added later.

Related story:
  https://www.pivotaltracker.com/story/show/22012217
  ("Accommodate projects without a set-up file")

Change-Id: I2dd8752e862f5c84949d3dc09b57ba8aa4928885
Reviewed-on: https://gerrit.causes.com/1722
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>