Don't start DRbServer on separate thread
Previously, we were explicitly starting the DRb server on a separate
thread. This appears to be unnecessary, as DRb.start_service spawns
a thread for us anyway.

This means we can remove our odd-looking condition variable control
flow code, which was originally used to prevent a race condition where
the server could be accessed before it was initialized.

Change-Id: I9cbb01172cf219a40a324b9aaaca64dd35cde9c3
Reviewed-on: https://gerrit.causes.com/1728
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>