Ensure .buffet directory exists on remote machines
Previously, Buffet would assume the existence of the .buffet directory
on the remote machine. This commit turns the existing `slave.execute`
method into `execute_in_project` (to emphasize that it `cd`s into the
project directory before executing a command) and adds a
`slave.execute` method for executing an arbitrary command on a remote
machine. This is then used to ensure the .buffet directory exists
in the home directory of the user on the remote machine.

Change-Id: If16d9abf365670024255a442a0582955a7afc954
Reviewed-on: https://gerrit.causes.com/1745
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>