Allow multiple users to run Buffet at the same time
If you try to run Buffet from two different machines for the same
project, there will be contention over who controls the project folder
for that project.

This commit fixes this problem by putting the project into a
subdirectory on a per-user basis. This way there is no overlap,
even if two different users are testing the same project.

It also passes a parameter to the prepare_script containing the user
name of the Buffet master. This allows the script to customize its
preparation based on who is running the test (e.g. using a different
database name generated from the Buffet user name) in order to
prevent contention.

Related story:
  http://www.pivotaltracker.com/story/show/22006583
  ("Allow multiple users to run Buffet at the same time")

Change-Id: I09bd791f7ea04fefcab0658799a3683c5d3180c2
Reviewed-on: https://gerrit.causes.com/1763
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>