Add --config switch to `buffet`
Defaulting to `buffet.yml` makes sense for most projects, and the file
should be checked into version control, but if you ever want to have a
config file with local tweaks and overrides, it would be nice to have a
way of getting Buffet to use it without overriding the official project
version.

This commit adds support for a custom config file to be specified with
the --config switch. It also updates the README to describe this new
capability.

Related story:
  http://www.pivotaltracker.com/story/show/22012835
  ("Add --config switch to `buffet`")

Change-Id: Ide5bbe2637bf090d77aa976dc68a59e0683c9aa0
Reviewed-on: https://gerrit.causes.com/1770
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>