Don't require project name/directory in buffet.yml
Buffet currently determines the location of the project directory it is
being run from using the buffet.yml file. This shouldn't be a required
setting, as it can easily determine the name/location of the project
directory using its current working directory (since that's where it
gets run from).

This commit removes the need for a project name/location to be
specified, inferring it from the current working directory.

Related story:
  http://www.pivotaltracker.com/story/show/22060491
  ("Don't require project name/directory in buffet.yml")

Change-Id: Iebf9805a86a5e4f6a3951f753f7f26deb0257d11
Reviewed-on: https://gerrit.causes.com/1774
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>