Don't wait for all slaves to run migrations before beginning tests
One simple optimization to make the overall process faster is to not
have all slaves wait for all other slaves to run the
bin/before-buffet-run script before running tests. Some machines are
faster than others, and it's a waste of time to force them to wait
around doing nothing.

This ends up improving the architecture a bit, as now the runner only
deals with starting the Master and displaying results, and the Master
is in charge of all operations done on a slave.

This change also adds logging to the CommandRunner for diagnostic
purposes.

Related story:
  http://www.pivotaltracker.com/story/show/22092541
  ("Don't wait for all slaves to run migrations before beginning tests")

Change-Id: If812a868f2c7ba634d74c3fe02f45a9c38c7951a
Reviewed-on: https://gerrit.causes.com/1788
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>