Teach Buffet to take --project option
Allow Buffet to use a specified project name for namespacing its
resources (and for passing on to the before-buffet-run script to
ensure the test databases used are unique for each project).

This will allow a user to run multiple tests at once for the same
project (before, you could only run multiple tests for different
projects, or more specifically, projects with different project
names).

Related story:
  http://www.pivotaltracker.com/story/show/22184455
  ("Teach Buffet to take --project option")

Change-Id: I2aca5f033fe8b6044df6a69baaaea82d8ec9e2f2
Reviewed-on: https://gerrit.causes.com/1841
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Shane da Silva <shane@causes.com>