Delete excluded project files on remote
While we weren't syncing files specified in the filter file, we were
also not deleting any old copies of these files if they already existed
on the remote. This could potentially have been bad if the log files
grew to exceptionally large sizes over time.

Adding the --delete-excluded flag to the rsync call takes care of this
problem for us.

Change-Id: I7e5026e8c485b65299b24726bcc390c70b1ed375
Reviewed-on: https://gerrit.causes.com/1854
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>