Add 'allowed_prepare_slave_failures' setting
This setting specifies how many slaves can fail in the prepare_slave
step without stopping the entire spec run, making it more resilient as
we increase the number of possible points of failure by adding more
machines.

At a high level, this commit:

  * Changes the messaging structure of the Master and its threads to use
    a ConditionVariable to signal/wait for updates from threads

  * Adds support for respecting the `allowed_prepare_slave_failures`
    setting to continue the test run unless the threshold has been
    exceeded

  * Updates the output from the Runner to display slave preparation
    failures

The bulk of this commit is in the redesigning of the Master to use a
condition variable as a signal mechanism. Previously, the Master spawned
a thread for each slave and then joined each of those threads so the
main thread would halt until all slaves had finished.

The problem with this approach is that if a failure occurs during
preparation, we don't find out about it until after we have joined the
failing thread. Thus a situation could arise where with 10 threads the
first 8 we spawn have no problems and run to completion, but the last 2
threads immediately fail due to being unable to connect to a host. In
this case, even though we could have stopped execution of the entire run
due to the 2 failures, we had to wait for the first 8 threads to finish.

The solution is to use signals so that the Master blocks on a single
condition variable, and can be signaled by any of the threads at any
time.

Change-Id: Id1852b4b7910c59a4262990447858c6a532ec703
Reviewed-on: https://gerrit.causes.com/16103
Reviewed-by: Lann Martin <lann@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Shane da Silva <shane@causes.com>