Show spec order for slaves with failures
In order to make debugging spurious test failures easier, we want to
keep track of the order in which a slave ran specs.

This adds a `specs` slave stat which is appended to each time a slave
takes a spec, and prints it out if a slave has any failures.

Change-Id: Ibf5a57a7ddb32ea93fffce4c5c3773b4dd739c77
Reviewed-on: https://gerrit.causes.com/27499
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>