Order specs by number of lines
This will hopefully speed up test runs by biasing specs which take
longer (assuming a heuristic that the number of lines in a spec is
indicative of how long that spec takes to run) to be run first. This in
theory means that we won't end up with any "slow stragglers" holding up
the test run, as all the specs that run at the end will be small.

Preliminary experiments show that test times drop from about 3 minutes
to 2.5 minutes, and there is also less variation in test times.

Change-Id: Ic5c16259717b2945dd07e4a8983ac93367e515a6
Reviewed-on: https://gerrit.causes.com/27828
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Tom Dooner <tom.dooner@causes.com>