Allow display of progress dots to be disabled
Add a `display_progress` option to `buffet.yml` which disables the
display of progress dots. This is useful for tools like Jenkins where
the dots are of no significant value.

This is a bit of a hack for now. In the future, we'll want to have a
bunch of formatters that can be specified instead of just
enabling/disabling dots.

Change-Id: Ia8856f80eccf77fefb37f284ff9e2673d309f654
Reviewed-on: https://gerrit.causes.com/33029
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>