Specify Buffet master/project with environment variables
Previously, Buffet would pass the master and project name via
command-line arguments to the `prepare_script`, but not the
`worker_command`.

In order to make working with these scripts consistent, and also to
allow the `worker_command` to have knowledge of the master and project,
change the mechanism to pass these arguments via environment variables.

Change-Id: I9b1d9570a985f6133500d3389c4624a4cdd43f55
Reviewed-on: https://gerrit.causes.com/33869
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>